:truck: Moved overlays and actionmenu
