+++ @navikt/core/react/package.json
@@ -71,12 +71,12 @@
     },
     "./ActionMenu": {
       "import": {
-        "types": "./esm/overlays/action-menu/index.d.ts",
-        "default": "./esm/overlays/action-menu/index.js"
+        "types": "./esm/action-menu/index.d.ts",
+        "default": "./esm/action-menu/index.js"
       },
       "require": {
-        "types": "./cjs/overlays/action-menu/index.d.ts",
-        "default": "./cjs/overlays/action-menu/index.js"
+        "types": "./cjs/action-menu/index.d.ts",
+        "default": "./cjs/action-menu/index.js"
       }
     },
     "./Alert": {
