+++ @navikt/core/react/src/action-menu/ActionMenu.stories.tsx
@@ -7,13 +7,13 @@ import {
   PencilIcon,
   StarIcon,
 } from "@navikt/aksel-icons";
-import { Button } from "../../button";
-import { InternalHeader } from "../../internal-header";
-import { Modal } from "../../modal";
-import { HStack, Spacer, VStack } from "../../primitives/stack";
-import { Theme } from "../../theme";
-import { Tooltip } from "../../tooltip";
-import { BodyShort, Detail } from "../../typography";
+import { Button } from "../button";
+import { InternalHeader } from "../internal-header";
+import { Modal } from "../modal";
+import { HStack, Spacer, VStack } from "../primitives/stack";
+import { Theme } from "../theme";
+import { Tooltip } from "../tooltip";
+import { BodyShort, Detail } from "../typography";
 import { ActionMenu } from "./ActionMenu";
 
 export default {
