+++ @navikt/core/react/src/action-menu/ActionMenu.tsx
@@ -1,16 +1,16 @@
 import React, { forwardRef, useRef } from "react";
 import { ChevronRightIcon } from "@navikt/aksel-icons";
-import { useModalContext } from "../../modal/Modal.context";
-import { type OverridableComponent, useId } from "../../utils-external";
-import { Slot } from "../../utils/components/slot/Slot";
+import { useModalContext } from "../modal/Modal.context";
+import { type OverridableComponent, useId } from "../utils-external";
+import { Menu, MenuPortalProps } from "../utils/components/floating-menu/Menu";
+import { Slot } from "../utils/components/slot/Slot";
 import {
   cl,
   composeEventHandlers,
   createStrictContext,
   requireReactElement,
-} from "../../utils/helpers";
-import { useControllableState, useMergeRefs } from "../../utils/hooks";
-import { Menu, MenuPortalProps } from "../floating-menu/Menu";
+} from "../utils/helpers";
+import { useControllableState, useMergeRefs } from "../utils/hooks";
 
 /* -------------------------------------------------------------------------- */
 /*                                 ActionMenu                                 */
