+++ @navikt/core/react/src/data/stories/Data.stories.tsx
@@ -10,8 +10,8 @@ import {
 } from "@tanstack/react-table";
 import React from "react";
 import { CogIcon, RectangleSectionsIcon } from "@navikt/aksel-icons";
+import { ActionMenu } from "../../action-menu";
 import { Button } from "../../button";
-import { ActionMenu } from "../../overlays/action-menu";
 import { VStack } from "../../primitives/stack";
 import DataActionBar from "../action-bar/root/DataActionBarRoot";
 import { DataTable } from "../table";
