+++ @navikt/core/react/src/data/table/th/DataTableTh.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from "react";
 import { FunnelIcon, SortDownIcon, SortUpIcon } from "@navikt/aksel-icons";
-import { ActionMenu } from "../../../overlays/action-menu";
+import { ActionMenu } from "../../../action-menu";
 import { HStack, Spacer } from "../../../primitives/stack";
 import { cl } from "../../../utils/helpers";
 import { DataTableThActions } from "./DataTableThActions";
