+++ @navikt/core/react/src/form/combobox/Combobox.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from "react";
-import { Floating } from "../../overlays/floating/Floating";
 import { BodyShort, ErrorMessage, Label } from "../../typography";
+import { Floating } from "../../utils/components/floating/Floating";
 import { cl } from "../../utils/helpers";
 import { ReadOnlyIconWithTitle } from "../ReadOnlyIcon";
 import ComboboxWrapper from "./ComboboxWrapper";
