+++ @navikt/core/react/src/form/combobox/Input/InputController.tsx
@@ -2,7 +2,7 @@
 
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 import React, { forwardRef } from "react";
-import { Floating } from "../../../overlays/floating/Floating";
+import { Floating } from "../../../utils/components/floating/Floating";
 import { cl } from "../../../utils/helpers";
 import { useMergeRefs } from "../../../utils/hooks";
 import { useFilteredOptionsContext } from "../FilteredOptions/filteredOptionsContext";
