+++ @navikt/core/react/src/form/file-upload/file-upload-item.stories.tsx
@@ -2,8 +2,8 @@ import { Meta, StoryFn } from "@storybook/react-vite";
 import React from "react";
 import { MenuElipsisVerticalCircleIcon } from "@navikt/aksel-icons";
 import { FileItem, FileUpload } from ".";
+import { ActionMenu } from "../../action-menu";
 import { Button } from "../../button";
-import { ActionMenu } from "../../overlays/action-menu";
 import { VStack } from "../../primitives/stack";
 
 const meta: Meta<typeof FileUpload.Item> = {
