+++ @navikt/core/react/src/utils/components/dismissablelayer/README.md
