+++ @navikt/core/react/src/utils/components/dismissablelayer/util/dispatchCustomEvent.ts
