+++ @navikt/core/react/src/utils/components/dismissablelayer/util/sort-layers.test.ts
