+++ @navikt/core/react/src/utils/components/dismissablelayer/util/sort-layers.ts
