+++ @navikt/core/react/src/utils/components/floating-menu/Menu.stories.tsx
