+++ @navikt/core/react/src/utils/components/floating/Floating.stories.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Portal } from "../../portal";
+import { Portal } from "../../../portal";
 import { Floating } from "./Floating";
 
 export default {
