+++ @navikt/core/react/src/utils/components/floating/Floating.utils.ts
