+++ @navikt/core/react/src/utils/components/focus-boundary/FocusBoundary.tests.stories.tsx
@@ -3,8 +3,8 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 import { act } from "@testing-library/react";
 import React, { useState } from "react";
 import { expect, fireEvent, userEvent, waitFor, within } from "storybook/test";
+import { ActionMenu } from "../../../action-menu";
 import { Dialog, DialogPopup } from "../../../dialog";
-import { ActionMenu } from "../../../overlays/action-menu";
 import { VStack } from "../../../primitives/stack";
 import { Provider } from "../../../provider";
 import { FocusBoundary, type FocusBoundaryProps } from "./FocusBoundary";
